moved suite test code to new file
